ClenshawCurtisQuadrature.jl Documentation

ClenshawCurtisQuadrature.interpolate — Method

interpolate(taus::AbstractVector{<:Real}, N::Integer; recursive::Bool = false)

Compute a matrix of Chebyshev polynomials of the first kind $T_n(\tau)$ for $\tau=taus$ and $n = 0,1,...,N$.

Arguments

taus::AbstractVector{<:Real}: The points at which to evaluate the Chebyshev polynomials.

N::Integer: The polynomial degree.

recursive::Bool: If true, use the recursive formula to compute the Chebyshev polynomials within the domain [-1,1]. If false, use the trigonometric formulation. Default is false.

Returns

Ts: A matrix of Chebyshev polynomial values at the given values of tau.

Description

This function computes the value of the N+1 Chebyshev polynomials at the given points in the domain [-1,1]. If the input taus is a scalar, it is converted to a vector. The function then computes the unweighted Chebyshev polynomial values at each tau.

ClenshawCurtisQuadrature.clenshaw_curtis_nested_ivpd — Method

clenshaw_curtis_nested_ivpd(N::Integer, M::Integer, d::Integer)

Compute the Clenshaw-Curtis quadrature and Chebyshev basis function matrices for a d-th order integral.

Arguments

N::Integer: The polynomial degree.

M::Integer: The sampling degree. Must bes greater than or equal to the polynomial degree. This is equal to the total number of function sampling points minus 1.

d::Integer: The integral order.

Returns

A: The Least Squares Operator matrix.

P: The Quadrature Matrix.

T: The Chebyshev Matrix.

Description

This function computes the Clenshaw-Curtis quadrature matrices and the basis function vectors a. It first generates the Chebyshev polynomials and the Least Squares Operator matrix using the lsq_chebyshev_fit function. Then, it calculates the Constants of Integration, and constructs the S matrix. Finally, it computes the Clenshaw Curtis Quadrature matrix.

ClenshawCurtisQuadrature.clenshaw_curtis_ivpi — Method

clenshaw_curtis_ivpi(N::Integer, M::Integer)

Compute the Clenshaw-Curtis quadrature and Cebyshev basis function matrices for a first order initial value problem.

Arguments

N::Integer: The polynomial degree.

M::Integer: The sampling degree. Must bes greater than or equal to the polynomial degree. This is equal to the total number of function sampling points minus 1.

Returns

A: The Least Squares Operator matrix.

Ta: The "acceleration" Chebyshev Matrix.

P1: The Quadrature Matrix for acceleration to velocity.

T1: The "Velocity" Chebyshev Matrix.

Description

This function computes the Clenshaw-Curtis quadrature matrices and the basis function vectors a. It first generates the Chebyshev polynomials and the Least Squares Operator matrix using the lsq_chebyshev_fit function. Then, it calculates the "Velocity" Constants of Integration, and constructs the S matrices for "velocity". Finally, it computes the Clenshaw Curtis Quadrature matrices for acceleration to velocity.

ClenshawCurtisQuadrature.clenshaw_curtis_ivpii — Method

clenshaw_curtis_ivpii(N::Integer, M::Integer)

Compute the Clenshaw-Curtis quadrature and Cebyshev basis function matrices for a second order initial value problem.

Arguments

N::Integer: The polynomial degree.

M::Integer: The sampling degree. Must bes greater than or equal to the polynomial degree. This is equal to the total number of function sampling points minus 1.

Returns

A: The Least Squares Operator matrix.

Ta: The "acceleration" Chebyshev Matrix.

P1: The Quadrature Matrix for acceleration to velocity.

T1: The "Velocity" Chebyshev Matrix.

P2: The Quadrature Matrix for velocity to position.

T2: The "Position" Chebyshev Matrix.

Description

This function computes the Clenshaw-Curtis quadrature matrices and the basis function vectors a. It first generates the Chebyshev polynomials and the Least Squares Operator matrix using the lsq_chebyshev_fit function. Then, it calculates the "Position" and "Velocity" Constants of Integration, and constructs the S matrices for "velocity" and "position". Finally, it computes the Clenshaw Curtis Quadrature matrices for acceleration to velocity and velocity to position.

Example

```julia N = 5 M = 5 A, Ta, P1, T1, P2, T2 = clenshawcurtisivpii(N, M)