ClenshawCurtisQuadrature.jl Documentation

Summary

ClenshawCurtisQuadrature.jl is a Julia package that provides tools for performing Clenshaw-Curtis quadrature, a numerical integration technique. This method is particularly useful for quickly integrating arbitrary functions with high accuracy. The package aims to offer efficient and easy-to-use functions for integration tasks.

Installation

ClenshawCurtisQuadrature.jl is currently in development and not included in any package registries. Users will need to manually add the package with this Github repository url

julia> Using Pkg; Pkg.add(url="https://github.com/DavidMSCode/ClenshawCurtisQuadrature.jl")

or enter the package manager with

julia> ]

and run

(@environment) pkg> add https://github.com/DavidMSCode/ClenshawCurtisQuadrature.jl

Usage

The purpose of this module is to numerically solve integrals of the form

\[F(\tau) = F(-1) + \int_{-1}^{\tau}f(\xi)d\xi\]

Note

If $F(\tau)$ is an antiderivative of $f(\tau)$, then $F(\tau)+c$ is also an antiderivative of $f(\tau)$ for any constant value $c$. But, if $F(-1)$ is known, then there is only one antiderivative that satisfies the integral.

where $f(\tau)$ is an integrand that is valid on the domain $[-1,1]$ and $F(\tau)$ is an antiderivative of the integrand. This is performed by representing $F(\tau)$ and $f(\tau)$ as a summation of Chebyshev polynomials on the domain $[-1,1]$

\[F(\tau) \approx \sum_{n=0}^N \alpha_n T_n(\tau)\]

\[f(\tau) \approx \sum_{n=0}^{N-1} a_n T_n(\tau)\]

where $T_n(\tau)$ is the $n$-th Chebyshev polynomial of the first kind. $N$ is the degree of the interpolating polynomial and $a_n$ and $\alpha_n$ are the interpolating coefficients of the integrand and the particular antiderivative respectively. Because the integrals of the Chebyshev polynomials are related the values of Chebyshev polynomials of higher and lower degrees, we can perform linear operations on the integrand's coefficients to calculate the corresponding antiderivative's coefficients. The functions in this module are designed to generate the matrices required for this quadrature.

Examples

If we want to know the integral on the domain $[-1,1]$ of the function $f(\tau)=\tau^2$, we can do so by first sampling the function at set of M+1 points. Because we know the antiderivative of $f(\tau)$ is a 3rd order polynomial, I'll choose $N=3$ as our desired interpolating polynomial of the solution. We can sample as many points as we want, but because the integrand will be represented as a 2nd order polynomial we need at least 3 points to calculate the coefficients.

To do so first generate the values of $\tau$ at cosine spaced nodes.

Note

Because we just wanted the value of the integral on the range $[-1,1]$ we assume that $F(-1)=0$. The next examples will demonstrating solving an initial value problem when $F(-1)\neq 0$.

The Ta and T1 matrices returned by clenshaw_curtis_ivpi() are also interpolating matrices. They specifically contain the values of the chebyshev polynomials needed to interpolate the integrand and antiderivative at each of the $M+1$ cosine spaced nodes we calculated before. Pre-calculating these matrices can be useful for iterative problems where the integrand is unkown as a function of $\tau$.

using ClenshawCurtisQuadrature

#set the degree variables
M = 2
N = 3

#Generate the cosine spaced nodes
Ms = 0:M
taus = cos.(pi*Ms/M)

#Sample the integrand
f = (x) -> x^2
ys = f.(taus)

#Get the Least Squares Operation and Quadrature matrix
A,Ta,P1,T1 = clenshaw_curtis_ivpi(N,M)

#Calculate the least squares fit interpolating polynomial coefficients
a = A*ys

#Apply the quadrature matrix
alpha = P1*a

#Get the interpolating polynomials at τ=1
T = interpolate(1,N)

F1 = T*alpha

# output

1-element Vector{Float64}:
 0.6666666666666666

The result is that $\int_{-1}^1 \tau^2d\tau=\frac{2}{3}$ which is exactly the correct answer.

Methods

ClenshawCurtisQuadrature.interpolate — Method

interpolate(τs::AbstractVector{<:Real}, N::Integer; recursive::Bool = false)

Computes a matrix of Chebyshev polynomials of the first kind $T_n(τ)$ for $τ ∈ τs$ and $n = 0,1,...,N$. Used with Chebyshev coefficients to calculate the value of the interpolating polynomials at the given values of $τ$.

Arguments

τs::AbstractVector{<:Real}: The points at which to evaluate the Chebyshev

polynomials.

N::Integer: The polynomial degree.

recursive::Bool: If true, use the recursive formula to compute the

Chebyshev polynomials within the domain [-1,1]. If false, use the trigonometric formulation. Default is false.

Returns

Ts: A matrix of Chebyshev polynomial values at the given values of τ.

ClenshawCurtisQuadrature.clenshaw_curtis_nested_ivpd — Method

clenshaw_curtis_nested_ivpd(d::Integer, N::Integer, M::Integer=N)

Compute the Clenshaw-Curtis quadrature and Chebyshev basis function matrices for an Nth degree polynomial and for a d-th order integral on the interval [-1,1].

Arguments

d::Integer: The integral order.

N::Integer: The polynomial degree.

M::Integer: The sampling degree. Must be greater than or equal to the

polynomial degree. The integrand is evaluated at M+1 cosine spaced nodes. Defaults to N.

Returns

A: The Least Squares Operator matrix.

P: The Quadrature Matrix.

T: The Chebyshev Matrix.

ClenshawCurtisQuadrature.clenshaw_curtis_ivpi — Method

clenshaw_curtis_ivpi(N::Integer, M::Integer=N)

Compute the Clenshaw-Curtis quadrature and Cebyshev basis function matrices for a first order initial value problem on the interval [-1,1].

Arguments

N::Integer: The polynomial degree.

M::Integer: The sampling degree. Must bes greater than or equal to the

polynomial degree. This is equal to the total number of function sampling points minus 1. Defaults to N.

Returns

A: The Least Squares Operator matrix.

Ta: The "acceleration" Chebyshev Matrix.

P1: The Quadrature Matrix for acceleration to velocity.

T1: The "Velocity" Chebyshev Matrix.

ClenshawCurtisQuadrature.clenshaw_curtis_ivpii — Method

clenshaw_curtis_ivpii(N::Integer, M::Integer=N)

Compute the Clenshaw-Curtis quadrature and Cebyshev basis function matrices for a second order initial value problem on the interval [-1,1].

Arguments

N::Integer: The polynomial degree.

M::Integer: The sampling degree. Must bes greater than or equal to the

polynomial degree. This is equal to the total number of function sampling points minus 1. Defaults to N.

Returns

A: The Least Squares Operator matrix.

Ta: The "acceleration" Chebyshev Matrix.

P1: The Quadrature Matrix for acceleration to velocity.

T1: The "Velocity" Chebyshev Matrix.

P2: The Quadrature Matrix for velocity to position.

T2: The "Position" Chebyshev Matrix.

Example

N = 5
M = 5
A, Ta, P1, T1, P2, T2 = clenshaw_curtis_ivpii(N, M)