ClenshawCurtisQuadrature.jl Documentation

Summary

ClenshawCurtisQuadrature.jl is a Julia package that provides tools for performing Clenshaw-Curtis quadrature, a numerical integration technique. This method is particularly useful for quickly integrating arbitrary functions with high accuracy. The package aims to offer efficient and easy-to-use functions for integration tasks.

Installation

ClenshawCurtisQuadrature.jl is currently in development and not included in any package registries. Users will need to manually add the package with this Github repository url

julia> Using Pkg; Pkg.add(url="https://github.com/DavidMSCode/ClenshawCurtisQuadrature.jl")

or enter the package manager with

julia> ]

and run

(@environment) pkg> add https://github.com/DavidMSCode/ClenshawCurtisQuadrature.jl

Usage

The purpose of this module is to numerically solve integrals of the form

\[F(\tau) = F(-1) + \int_{-1}^{1}f(\tau)d\tau\]

where $f(\tau)$ is an integrand that is valid on the domain [-1,1] and $F(\tau)$ is an antiderivative of the integrand.

Note

If $F(\tau)$ is an antiderivative of $f(\tau)$, then $F(\tau)+c$ is also an antiderivative of $f(\tau)$ for any constant value $c$. But, if $F(-1)$ is known, then there is only one antiderivative that satisfies the integral.

Methods

ClenshawCurtisQuadrature.interpolate — Method

interpolate(τs::AbstractVector{<:Real}, N::Integer; recursive::Bool = false)

Computes a matrix of Chebyshev polynomials of the first kind $T_n(τ)$ for $τ ∈ τs$ and $n = 0,1,...,N$. Used with Chebyshev coefficients to calculate the value of the interpolating polynomials at the given values of $τ$.

Arguments

τs::AbstractVector{<:Real}: The points at which to evaluate the Chebyshev

polynomials.

N::Integer: The polynomial degree.

recursive::Bool: If true, use the recursive formula to compute the

Chebyshev polynomials within the domain [-1,1]. If false, use the trigonometric formulation. Default is false.

Returns

Ts: A matrix of Chebyshev polynomial values at the given values of τ.

ClenshawCurtisQuadrature.clenshaw_curtis_nested_ivpd — Method

clenshaw_curtis_nested_ivpd(d::Integer, N::Integer, M::Integer=N)

Compute the Clenshaw-Curtis quadrature and Chebyshev basis function matrices for an Nth degree polynomial and for a d-th order integral on the interval [-1,1].

Arguments

d::Integer: The integral order.

N::Integer: The polynomial degree.

M::Integer: The sampling degree. Must be greater than or equal to the

polynomial degree. The integrand is evaluated at M+1 cosine spaced nodes. Defaults to N.

Returns

A: The Least Squares Operator matrix.

P: The Quadrature Matrix.

T: The Chebyshev Matrix.

ClenshawCurtisQuadrature.clenshaw_curtis_ivpi — Method

clenshaw_curtis_ivpi(N::Integer, M::Integer=N)

Compute the Clenshaw-Curtis quadrature and Cebyshev basis function matrices for a first order initial value problem on the interval [-1,1].

Arguments

N::Integer: The polynomial degree.

M::Integer: The sampling degree. Must bes greater than or equal to the

polynomial degree. This is equal to the total number of function sampling points minus 1. Defaults to N.

Returns

A: The Least Squares Operator matrix.

Ta: The "acceleration" Chebyshev Matrix.

P1: The Quadrature Matrix for acceleration to velocity.

T1: The "Velocity" Chebyshev Matrix.

ClenshawCurtisQuadrature.clenshaw_curtis_ivpii — Method

clenshaw_curtis_ivpii(N::Integer, M::Integer=N)

Compute the Clenshaw-Curtis quadrature and Cebyshev basis function matrices for a second order initial value problem on the interval [-1,1].

Arguments

N::Integer: The polynomial degree.

M::Integer: The sampling degree. Must bes greater than or equal to the

polynomial degree. This is equal to the total number of function sampling points minus 1. Defaults to N.

Returns

A: The Least Squares Operator matrix.

Ta: The "acceleration" Chebyshev Matrix.

P1: The Quadrature Matrix for acceleration to velocity.

T1: The "Velocity" Chebyshev Matrix.

P2: The Quadrature Matrix for velocity to position.

T2: The "Position" Chebyshev Matrix.

Example

N = 5
M = 5
A, Ta, P1, T1, P2, T2 = clenshaw_curtis_ivpii(N, M)